v2.6.0

+++ include/asm-ppc/hdreg.h
@@ -1,5 +1,5 @@
 /*
- *  linux/include/asm-ppc/hdreg.h
+ *  include/asm-ppc/hdreg.h
  *
  *  Copyright (C) 1994-1996  Linus Torvalds & authors
  */
@@ -11,7 +11,7 @@
 #ifndef __ASMPPC_HDREG_H
 #define __ASMPPC_HDREG_H
 
-typedef unsigned int ide_ioreg_t;
+typedef unsigned long ide_ioreg_t;
 
 #endif /* __ASMPPC_HDREG_H */
